tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/060.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: appending code to initially non-empty SCRIPT tag in DOM after setting innerText/innerHTML</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
	<script>log('HEAD script');</script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		var script=document.getElementsByTagName('script')[3];
		script.innerHTML='';
		script.appendChild( document.createTextNode('log("injected script code 1");') );
		script.innerText='';
		script.appendChild( document.createTextNode('log("injected script code 2");') );
		log('end script #1');
	</script>
	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_array_equals(eventOrder, ['HEAD script', 'inline script #1', 'end script #1', 'inline script #2']);
		t.done();
}
	onload = t.step_func(test);
	</script>

</body></html>
